Simplify GitHub workflow for Homebrew
This change simplifies and modifies the GitHub workflow set up for Homebrew. The matrix strategy that was previously implemented has been removed. It adjusted the download links for the release tarball and updated the Ruby file accordingly. Search for a specific file ('asd.rb') was also added whilst commit and push mechanism was simplified.